Shared Lib - Created Noise notifier trait Added a trait that sends a notification that a noise was heard by another robot to montroyashi
