+++ poutine_production/nordo_service/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2018"
 [dependencies]
 serde = { version = "1.0.105", features = ["derive"] }
 serde_json = "1.0.50"
-shared_models = {path = "../shared_models"}
+shared = {path = "../shared"}
 futures = "0.3.5"
 rand = "0.7.3"
 tokio = { version = "0.2.22", features = ["full"] }
