+++ poutine_production/shared/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "shared_models"
+name = "shared"
 version = "0.1.0"
 authors = ["Pierre <pseguin2011@protonmail.com>"]
 edition = "2018"
@@ -8,3 +8,4 @@ edition = "2018"
 
 [dependencies]
 serde = { version = "1.0.105", features = ["derive"] }
+easy-http-request = "0.2.9"
