--- poutine_production/shared_models/src/lib.rs
@@ -1,17 +1,0 @@
-use serde::{Deserialize, Serialize};
-/// Shared Models between multiple services
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Potato {
-    pub size: usize,
-    pub coated_in_maple_syrup: bool,
-    pub boiled: bool,
-    pub fried: bool,
-    pub oil_used: Option<OilTypes>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum OilTypes {
-    Sunflower,
-    Canola,
-    Vegetable,
-}
